requirements.txt
aiogram
coloredlogs
toml
pillow